requirements.txt
mysql-connector-python
requests
discord